mod: added stamina values and changed the inputs for player movement
The player movement was relying on set keybinds, and I chagned it to work with the keybind manager, but errors appears and I don't think the script was ready yet so I set it back to normal. I also changed the stamina values so that the current value wasn't at 1, which didn't let the player be able to sprint. I set the current and max value to 100 so the player can sprint around/ There is still an error where the player cannot move at all when sprinting.