mod: Created empty objects for camera control script and attached script to player
The camera control script has to have a First Person Snap, a Player, a main camera, a Third Person Parent and a Third Person Snap. Those objects were not set up properly and the script was not attached to the player. I have gone ahead and changed that, and the camera control should now be working as intended